Fix comparison between signed and unsigned integer expressions

+++ lib/Microphone/Microphone.cpp
@@ -16,7 +16,7 @@ int Microphone::read() const {
 
 bool Microphone::is_triggered(unsigned short treshold, unsigned short streak) {
     int reading = this->read();
-    if (reading == treshold) {
+    if (reading == static_cast<int>(treshold)) {
         this->trigger_streak += 1;
         if (this->trigger_streak == streak) {
             this->trigger_streak = 0;
